+++ djstripe/management/commands/djstripe_init_plans.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+from django.conf import settings
 from django.core.management.base import BaseCommand
 
 from ...sync import sync_plans