+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
 * SubscribeFormView and ChangePlanView no longer populate self.error with form errors
 * InvoiceItems.plan can now be null (as it is with individual charges), resolving #140 (Thanks @awechsler and @MichelleGlauser for help troubleshooting)
 * Email templates are now packaged during distribution.
+* sync_plans now takes an optional api_key
 
 0.5.0 (2015-05-25)
 ---------------------